+++ Palette.qml
@@ -32,4 +32,6 @@ PaletteForm {
         onClicked: paletteBack.ApplyChanges()
     }
 
+    testButton.onClicked: paletteBack.testCustomPalette()
+
 }