+++ PaletteForm.ui.qml
@@ -6,6 +6,7 @@ Page {
     width: 450
     height: 800
     anchors.fill: parent
+    property alias testButton: testButton
     property alias solidPaletteSwitch: solidPaletteSwitch
     property alias brightnessSlider: brightnessSlider
     property alias delayBox: delayBox
@@ -36,7 +37,7 @@ Page {
         currentIndex: 0
         anchors.verticalCenterOffset: -75
 
-        model: ["Rainbow", "Rainbow Stripe", "Cloud", "Party", "Ocean", "Lava", "Forest"]
+        model: ["Custom", "Rainbow", "Rainbow Stripe", "Cloud", "Party", "Ocean", "Lava", "Forest"]
     }
 
     SwitchDelegate {
@@ -114,5 +115,12 @@ Page {
         anchors.verticalCenter: parent.verticalCenter
         anchors.horizontalCenter: parent.horizontalCenter
         anchors.verticalCenterOffset: -25
+    }
+
+    Button {
+        id: testButton
+        x: 191
+        y: 231
+        text: qsTr("Test")
     }
 }